obs slice int

+++ agents/pearl/replay_buffer.py
@@ -93,7 +93,8 @@ class PEARLReplayBuffer(AbstractOnlineReplayBuffer):
             size=batch_size,
         )
         observation_slice = [
-            np.arange(idx - self.history_length, idx + 0.1) for idx in sample_indices
+            np.arange(idx - self.history_length, idx + 0.1, dtype=int)
+            for idx in sample_indices
         ]  # 0.1 to include idx
 
         # model inputs
